move session route into own file
